requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
cfgv==3.3.1
distlib==0.3.6
Django==4.1.3
django-debug-toolbar==3.7.0
django-decouple==2.1
django-ninja==1.1.0
filelock==3.12.2
identify==2.5.24
nodeenv==1.8.0
platformdirs==2.5.4
pre-commit==2.20.0
pydantic==2.5.3
pydantic_core==2.14.6
PyYAML==6.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
virtualenv==20.21.1
